refactor(cli): isolate Todo registration and Turn writeback adapters
