+++ tests/control_plane/test_split_root_todo_writeback_fence.py
@@ -14,7 +14,7 @@ from typing import Any
 
 import pytest
 
-from loopx.cli_commands.turn import (
+from loopx.cli_commands.turn_todo_writeback import (
     write_turn_repair_update,
     write_turn_validated_completion,
 )
